Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

DerivativeFreeSolvers.coordinate_search

DerivativeFreeSolvers.mads

DerivativeFreeSolvers.nelder_mead

​

DerivativeFreeSolvers.coordinate_search — Method

coordinate_search(nlp)

The coordinate search is a derivative free optimization method. It minimizes a certain function f by walking on a grid and taking a step towards smaller function values. When it can no longer find any smaller values, it reduces the grid by a factor of β, repeating the process.

DerivativeFreeSolvers.mads — Method

mads(nlp)

MADS is implemented following the description of

Mark A. Abramson, Charles Audet, J. E. Dennis Jr., and Sébastien Le Digabel.
OrthoMADS: A Deterministic MADS Instance with Orthogonal Directions.
SIAM Journal on Optimization, 2009, Vol. 20, No. 2, pp. 948-966,

But using random vectors instead of Halton vectors, as suggested by Francisco Sobral.

DerivativeFreeSolvers.nelder_mead — Method

nelder_mead(nlp)

This implementation follows the algorithm described in chapter 9 of [1]. The Oriented Restart follows [2]. [1] Numerical Optimization (Jorge Nocedal and Stephen J. Wright), Springer, 2006. [2] C. T. Kelley. Detection and remediation of stagnation in the nelder–mead algorithm using a sufficient decrease condition. SIAM J. on Optimization, 1999.